requirements.txt
altair==5.0.1
asn1crypto
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Bottleneck
brotlipy==0.7.0
cachetools==5.3.1
certifi
cffi
charset-normalizer
click==8.1.3
cloudpickle
colorama==0.4.6
contourpy==1.0.7
cryptography
cycler==0.11.0
decorator==5.1.1
extra-streamlit-components==0.1.56
Faker==18.10.1
favicon==0.7.0
filelock
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
hasher==2.0.1
htbuilder==0.6.1
idna
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
# mkl-fft==1.3.1
# mkl-random
# mkl-service==2.4.0
more-itertools==9.1.0
networkx==3.1
numexpr
numpy
oscrypto
packaging
pandas
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.23.2
pyarrow==10.0.1
pycparser
pycryptodomex
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT
pymdown-extensions==10.0.1
Pympler==1.0.1
pyOpenSSL 
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks
python-dateutil 
pytz 
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests
rich==13.4.1
six
smmap==5.0.0
snowflake-connector-python
snowflake-snowpark-python 
soupsieve==2.4.1
st-annotated-text==4.0.0
streamlit==1.23.1
streamlit-authenticator==0.2.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-chat==0.0.2.2
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions
tzdata==2023.3
tzlocal==4.3
Unidecode==1.3.6
urllib3
validators==0.20.0
watchdog==3.0.0
win-inet-pton
zipp==3.15.0